Menabahkan Role/Permission Middleware dan Menambahkan Role pada Operasio CRUD User

+++ routes/api.php
@@ -8,8 +8,8 @@ Route::middleware(['auth:sanctum'])->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::middleware(['auth:sanctum'])->get('/users', [UserController::class, 'index']);
-Route::middleware(['auth:sanctum'])->get('/users/{user}', [UserController::class, 'show']);
-Route::middleware(['auth:sanctum'])->post('/users', [UserController::class, 'store']);
-Route::middleware(['auth:sanctum'])->patch('/users/{user}', [UserController::class, 'update']);
-Route::middleware(['auth:sanctum'])->delete('/users/{user}', [UserController::class, 'destroy']);
+Route::middleware(['auth:sanctum', 'role:Admin'])->get('/users', [UserController::class, 'index']);
+Route::middleware(['auth:sanctum', 'role:Admin'])->get('/users/{user}', [UserController::class, 'show']);
+Route::middleware(['auth:sanctum', 'role:Admin'])->post('/users', [UserController::class, 'store']);
+Route::middleware(['auth:sanctum', 'role:Admin'])->patch('/users/{user}', [UserController::class, 'update']);
+Route::middleware(['auth:sanctum', 'role:Admin'])->delete('/users/{user}', [UserController::class, 'destroy']);
